play a little song

+++ testing/Keys.hs
@@ -3,5 +3,5 @@ module Keys (keyToFreq) where
 -- this is an awful makeshift solution
 -- maybe we could use something similar to Haskore.Basic.Pitch
 
-keyToFreq :: Double -> Double
-keyToFreq k = 440.0*2**((k-49.0)/12)
+keyToFreq :: Int -> Double
+keyToFreq k = 440.0*2**(((fromIntegral k)-49.0)/12)
